+++ Omise.Net/Services/ChargeService.cs
@@ -18,12 +18,33 @@ namespace Omise
         }
 
         /// <summary>
+        /// Initializes a new instance of the <see cref="Omise.ChargeService"/> class with Api key. The service uses default request manager object.
+        /// </summary>
+        /// <param name="apiKey">Api key</param>
+        /// <param name="apiVersion">Api version</param>
+        public ChargeService(string apiKey, string apiVersion)
+            : base(apiKey, apiVersion)
+        {
+        }
+
+        /// <summary>
         /// Initializes a new instance of the <see cref="Omise.ChargeService"/> class with IRequestManager object and Api key
         /// </summary>
         /// <param name="requestManager">IRequestManager object</param>
         /// <param name="apiKey">Api key</param>
         public ChargeService(IRequestManager requestManager, string apiKey)
             : base(requestManager, apiKey)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="Omise.ChargeService"/> class with IRequestManager object and Api key
+        /// </summary>
+        /// <param name="requestManager">IRequestManager object</param>
+        /// <param name="apiKey">Api key</param>
+        /// <param name="apiVersion">Api version</param>
+        public ChargeService(IRequestManager requestManager, string apiKey, string apiVersion)
+            : base(requestManager, apiKey, apiVersion)
         {
         }