+++ Omise.Net/Services/ServiceBase.cs
@@ -80,21 +80,8 @@ namespace Omise
             get { return "https://api.omise.co"; }
         }
 
-        /// <summary>
-        /// Initializes the ServiceBase object
-        /// </summary>
-        /// <param name="requestManager">IRequestManager object</param>
-        /// <param name="apiKey">Api key</param>
-        public ServiceBase(IRequestManager requestManager, string apiKey)
+        private void init()
         {
-            if (requestManager == null)
-            {
-                requester = new RequestManager(ApiUrlBase, apiKey);
-            }
-            else
-            {
-                requester = requestManager;
-            }
             accountFactory = new AccountFactory();
             balanceFactory = new BalanceFactory();
             bankAccountFactory = new BankAccountFactory();
@@ -112,21 +99,50 @@ namespace Omise
         /// <summary>
         /// Initializes the ServiceBase object
         /// </summary>
+        /// <param name="requestManager">IRequestManager object</param>
+        /// <param name="apiKey">Api key</param>
+        public ServiceBase(IRequestManager requestManager, string apiKey)
+        {
+            if (requestManager == null)
+            {
+                requester = new RequestManager(ApiUrlBase, apiKey);
+            }
+            else
+            {
+                requester = requestManager;
+            }
+
+            init();
+        }
+
+        /// <summary>
+        /// Initializes the ServiceBase object
+        /// </summary>
         /// <param name="apiKey">Api key</param>
         public ServiceBase(string apiKey)
         {
             requester = new RequestManager(ApiUrlBase, apiKey);
-            accountFactory = new AccountFactory();
-            balanceFactory = new BalanceFactory();
-            bankAccountFactory = new BankAccountFactory();
-            cardFactory = new CardFactory();
-            chargeFactory = new ChargeFactory();
-            customerFactory = new CustomerFactory();
-            recipientFactory = new RecipientFactory();
-            refundFactory = new RefundFactory();
-            tokenFactory = new TokenFactory();
-            transferFactory = new TransferFactory();
-            disputeFactory = new DisputeFactory();
+            init();
+        }
+
+        public ServiceBase(string apiKey, string apiVersion)
+        {
+            requester = new RequestManager(ApiUrlBase, apiKey, apiVersion);
+            init();
+        }
+
+        public ServiceBase(IRequestManager requestManager, string apiKey, string apiVersion)
+        {
+            if (requestManager == null)
+            {
+                requester = new RequestManager(ApiUrlBase, apiKey, apiVersion);
+            }
+            else
+            {
+                requester = requestManager;
+            }
+
+            init();
         }
 
         /// <summary>