+++ Omise.Net/Services/TransferService.cs
@@ -29,6 +29,27 @@ namespace Omise
         }
 
         /// <summary>
+        /// Initializes a new instance of the <see cref="Omise.TransferService"/> class with Api key. The service uses default IRequestManager object.
+        /// </summary>
+        /// <param name="apiKey">Api key</param>
+        /// <param name="apiVersion">Api version</param>
+        public TransferService(string apiKey, string apiVersion)
+            : base(apiKey, apiVersion)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="Omise.TransferService"/> class with IRequestManager object and Api key
+        /// </summary>
+        /// <param name="requestManager">IRequestManager object</param>
+        /// <param name="apiKey">Api key</param>
+        /// <param name="apiVersion">Api version</param>
+        public TransferService(IRequestManager requestManager, string apiKey, string apiVersion)
+            : base(requestManager, apiKey, apiVersion)
+        {
+        }
+
+        /// <summary>
         /// Gets all transfers.
         /// </summary>
         /// <returns>CollectionResponseObject of transfers.</returns>